feat: Network Security Config
This allows the application to access the unencrypted backend application. The network_security_config.xml file allows our application to be able to communicate with the IP address 10.0.2.2 which is the address of the web server that our locally ran Spring Boot API application will be running on. This configuration allows communication with that, so we can make HTTP requests to the server.