feat: creates a method that adds an album to the album repository
The enqueue method in call is used to initiate an asynchronous HTTP request and get the response from the server, typically to execute the request on a background thread, which is internally managed by Retrofit. The response is delivered on the main thread. The Callback is invoked when the request is completed.